Add docker check before prune

+++ WinChad.ps1
@@ -121,4 +121,4 @@ Set-FTA Applications\notepad++.exe .txt
 
 # cleanup
 choco upgrade choco-cleaner --params "'/NOTASK:TRUE'" -y --ignore-dependencies; Start-Process -FilePath "C:\ProgramData\chocolatey\bin\choco-cleaner.bat" -Wait
-docker system prune --volumes -f
+if (docker info > $null 2>&1) { docker system prune --volumes -f }
